+++ app/src/main/AndroidManifest.xml
@@ -7,14 +7,14 @@
     <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE"/>
 
     <application
-        android:name=".example.BootstrapApplication"
+        android:name=".training.BootstrapApplication"
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:theme="@style/AppTheme">
 
         <activity
-            android:name=".example.ui.example.ExampleActivity"
+            android:name=".training.ui.login.LoginActivity"
             android:screenOrientation="portrait">
 
             <intent-filter>
@@ -23,10 +23,6 @@
             </intent-filter>
 
         </activity>
-
-        <activity
-            android:name=".example.ui.viewpager.ViewpagerActivity"
-            android:screenOrientation="portrait" />
 
     </application>
