--- app/src/main/java/ar/com/wolox/android/example/BaseConfiguration.kt
@@ -1,10 +1,0 @@
-package ar.com.wolox.android.example
-
-internal open class BaseConfiguration {
-
-    companion object {
-        const val EXAMPLE_CONFIGURATION_KEY = "https://jsonplaceholder.typicode.com"
-        const val SHARED_PREFERENCES_NAME = "private-shared-prefs"
-    }
-
-}
