--- app/src/main/java/ar/com/wolox/android/example/di/module/MiscModule.kt
@@ -1,14 +1,0 @@
-package ar.com.wolox.android.example.di.module
-
-import ar.com.wolox.wolmo.core.presenter.BasePresenter
-import dagger.Module
-import dagger.Provides
-
-@Module
-class MiscModule {
-
-    @Provides
-    internal fun provideBasePresenter(): BasePresenter<Any> {
-        return BasePresenter()
-    }
-}
