--- app/src/main/java/ar/com/wolox/android/example/model/ExampleModel.kt
@@ -1,3 +1,0 @@
-package ar.com.wolox.android.example.model
-
-data class ExampleModel(val someNumber: Int)
