--- app/src/main/java/ar/com/wolox/android/example/model/Post.kt
@@ -1,3 +1,0 @@
-package ar.com.wolox.android.example.model
-
-data class Post(val title: String, val body: String)
