--- app/src/main/java/ar/com/wolox/android/example/model/ViewPagerModel.kt
@@ -1,3 +1,0 @@
-package ar.com.wolox.android.example.model
-
-data class ViewPagerModel(val name: String)
