--- app/src/main/java/ar/com/wolox/android/example/ui/example/IExampleView.kt
@@ -1,8 +1,0 @@
-package ar.com.wolox.android.example.ui.example
-
-interface IExampleView {
-
-    fun onUsernameSaved()
-
-    fun onSignUp()
-}
