--- app/src/main/java/ar/com/wolox/android/example/ui/viewpager/random/IRandomView.kt
@@ -1,8 +1,0 @@
-package ar.com.wolox.android.example.ui.viewpager.random
-
-interface IRandomView {
-
-    fun setUsername(username: String)
-
-    fun onRandomNumberUpdate(someNumber: Int)
-}
