--- app/src/main/java/ar/com/wolox/android/example/ui/viewpager/request/IRequestView.kt
@@ -1,10 +1,0 @@
-package ar.com.wolox.android.example.ui.viewpager.request
-
-interface IRequestView {
-
-    fun setNewsTitle(title: String)
-
-    fun setNewsBody(body: String)
-
-    fun showError()
-}
