+++ app/src/main/java/ar/com/wolox/android/training/BaseConfiguration.kt
@@ -1,0 +1,10 @@
+package ar.com.wolox.android.training
+
+internal open class BaseConfiguration {
+
+    companion object {
+        const val EXAMPLE_CONFIGURATION_KEY = "https://jsonplaceholder.typicode.com"
+        const val SHARED_PREFERENCES_NAME = "private-shared-prefs"
+    }
+
+}
