+++ app/src/main/java/ar/com/wolox/android/training/model/ExampleModel.kt
@@ -1,0 +1,3 @@
+package ar.com.wolox.android.training.model
+
+data class ExampleModel(val someNumber: Int)
