+++ app/src/main/java/ar/com/wolox/android/training/model/Post.kt
@@ -1,0 +1,3 @@
+package ar.com.wolox.android.training.model
+
+data class Post(val title: String, val body: String)
