+++ app/src/main/java/ar/com/wolox/android/training/model/ViewPagerModel.kt
@@ -1,0 +1,3 @@
+package ar.com.wolox.android.training.model
+
+data class ViewPagerModel(val name: String)
