+++ app/src/main/java/ar/com/wolox/android/training/ui/home/HomeActivity.kt
@@ -1,0 +1,14 @@
+package ar.com.wolox.android.training.ui.home
+
+import ar.com.wolox.android.R
+import ar.com.wolox.wolmo.core.activity.WolmoActivity
+import javax.inject.Inject
+
+class HomeActivity @Inject constructor() : WolmoActivity() {
+
+    override fun layout(): Int = R.layout.activity_base
+
+    override fun init() {
+        replaceFragment(R.id.vActivityBaseContent, HomeFragment())
+    }
+}
