+++ app/src/main/java/ar/com/wolox/android/training/ui/login/ILoginView.kt
@@ -1,0 +1,6 @@
+package ar.com.wolox.android.training.ui.login
+
+interface ILoginView {
+
+    fun onUsernameSaved()
+}
