+++ app/src/main/java/ar/com/wolox/android/training/ui/login/LoginPresenter.kt
@@ -1,0 +1,13 @@
+package ar.com.wolox.android.training.ui.login
+
+import ar.com.wolox.android.training.utils.UserSession
+import ar.com.wolox.wolmo.core.presenter.BasePresenter
+import javax.inject.Inject
+
+class LoginPresenter @Inject constructor(private val mUserSession: UserSession) : BasePresenter<ILoginView>(){
+
+    fun storeUsername(text: String) {
+        mUserSession.username = text
+        view.onUsernameSaved()
+    }
+}
