+++ app/src/main/java/ar/com/wolox/android/training/utils/Extras.kt
@@ -1,0 +1,16 @@
+package ar.com.wolox.android.training.utils
+
+import android.app.Activity
+import android.content.SharedPreferences
+import android.support.v4.app.Fragment
+
+/**
+ * Util class to store keys to use with [SharedPreferences] or as argument between
+ * [Fragment] or [Activity].
+ */
+object Extras {
+
+    object UserLogin {
+        const val USERNAME = "username"
+    }
+}
