+++ app/src/main/res/layout/fragment_login.xml
@@ -62,6 +62,7 @@
         android:layout_height="wrap_content"
         android:layout_marginEnd="16dp"
         android:layout_marginStart="16dp"
+        android:onClick="LogIn"
         android:text="@string/fragment_example_login"
         app:layout_constraintBottom_toTopOf="@+id/vSignUpButton"
         app:layout_constraintEnd_toEndOf="parent"
@@ -76,6 +77,7 @@
         android:layout_marginBottom="18dp"
         android:layout_marginEnd="16dp"
         android:layout_marginStart="16dp"
+        android:onClick="SignUp"
         android:backgroundTint="@color/chuck_status_300"
         android:text="@string/fragment_example_signup"
         android:textColor="@color/chuck_status_requested"
