--- app/src/main/res/layout/fragment_request.xml
@@ -1,50 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<LinearLayout
-    xmlns:android="http://schemas.android.com/apk/res/android"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical">
-
-    <android.support.v7.widget.Toolbar
-        android:id="@+id/vToolbar"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:background="@color/colorPrimary" />
-
-    <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:padding="@dimen/spacing_large"
-        android:gravity="center"
-        android:orientation="vertical">
-
-        <TextView
-            android:id="@+id/vPageTitle"
-            android:layout_width="match_parent"
-            android:layout_height="@dimen/spacing_none"
-            android:layout_weight="1"
-            android:layout_margin="@dimen/spacing_medium_more"
-            android:gravity="top|center_horizontal"
-            android:text="@string/page_2_title"
-            android:textSize="@dimen/font_big" />
-
-        <TextView
-            android:id="@+id/vPageBody"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:layout_margin="@dimen/spacing_medium_more"
-            android:gravity="center"
-            android:text="@string/example_textview"
-            android:textSize="@dimen/font_medium_extra" />
-
-        <TextView
-            android:layout_width="match_parent"
-            android:layout_height="@dimen/spacing_none"
-            android:layout_weight="1"
-            android:layout_margin="@dimen/spacing_medium_more"
-            android:gravity="bottom|center_horizontal"
-            android:text="@string/example_wolox"
-            android:textSize="@dimen/font_medium_extra" />
-
-    </LinearLayout>
-</LinearLayout>
