--- app/src/production/java/ar/com/wolox/android/example/Configuration.kt
@@ -1,7 +1,0 @@
-package ar.com.wolox.android.example
-
-internal class Configuration : BaseConfiguration() {
-    companion object {
-        const val API_ENDPOINT = "http://api.wolox.com.ar"
-    }
-}
